+++ hooks/__tests__/useCardCompanySelect.test.ts
@@ -1,6 +1,6 @@
 import { renderHook, act } from '@testing-library/react';
 import { useCardCompanySelect } from '../src/lib/useCardCompanySelect';
-import { CardCompany } from '../src/lib/types/card';
+import { CardCompany } from '../src/lib/card.type';
 
 describe('useCardCompanySelect custom hook 테스트', () => {
   test.each(['BC카드', '신한카드', '카카오뱅크', '현대카드', '우리카드', '롯데카드', '하나카드', '국민카드'])(
